% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLms.R
\name{runLms}
\alias{runLms}
\title{Run lm() ROI stats in a linear model, per your specifications}
\usage{
runLms(df, regressionString, lut)
}
\arguments{
\item{df}{A data frame in long format, organized as QuANTs-style output: dependent variable (freq. thicknesses, volumes, or cbf values, etc) in a column called "value"
a column called "label" with the label numbers. Each row is one measurement for one person. Can have as many columns as you want (ie, ignores everything not specified")
Also requires columns named as specified  used as varOfInterest, covarOfInterest (if used), and otherCovarString (if used)}

\item{regressionString}{An "lm-compatible" string of what you want your regression to be, using the column names of your df 
eg, "value~PatientGroup+AgeAtMRI+PatientGroup:AgeAtMRI"}

\item{lut}{A look up table with label numbers in the first column and label names in the second column. Can have extra rows or columns, they will be ignored}
}
\value{
A list object with 3 elements 
    1) a df containing significant labels from the lm()
    2) summay(lm()) output from each label in your input df 
    3) a plot for each region in your df 1) (ie, one plot for each significant label)
}
\description{
Run lm() ROI stats in a linear model, per your specifications
}
