% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLmsPlots.R
\name{runLmsPlots}
\alias{runLmsPlots}
\title{Make a plot of a bunch of ROI stats from a linear model, per your specifications}
\usage{
runLmsPlots(
  plotList,
  termToPlot,
  x,
  y,
  interactionPlot = FALSE,
  boxOrScatter = "box",
  colorCol,
  colorMat = NULL,
  sigValue = 1,
  multipleCorrectionMethod = "none",
  outFileRoot = NULL
)
}
\arguments{
\item{plotList}{A list object with 1st element a stats table organized as long in terms of Label.Names x term (from your lm output)
The 2nd element is a data frame with columns "value", "label", "Label.Names", "resid", plus whatever you stuck into your lm()
The width in the df should be organized as estimate, std.error, statistic, and p.value
Needs raw values ("value"), as well as columns associated with each term in your regression
Note that you can do this with only a single row, or a few select rows, but then don't use the function's multiple corrections tests}

\item{x}{what you want on your x-axis. If it's a factor or character, this makes a boxplot. If it's a numeric, double, or integer, it's a scatterplot.}

\item{y}{what you want on your y-axis. Almost certainly "value" or "resid"}

\item{interactionPlot}{(logical) are you plotting an interaction effect? Should only be used when considering an interaction effect statistically. Default FALSE}

\item{boxOrScatter}{"box" or "scatter" depending on how you want to visualize your data}

\item{colorCol}{what you want to color by}

\item{colorMat}{(optional) table of colors that might get used in plots. If used, might fight with colorCol. I need to practice}

\item{sigValue}{value used for thresholding your statistics. If you want to show all data use 1 (default).}

\item{multipleCorrectionMethod}{(optional) goes into p.adjust() to correct for multiple comparisons. Will do this accounting for all labels in your input df. Default is none, common alternatives are "bonferroni" and "fdr"}

\item{outFileRoot}{(optional) Not implemented. file root for output. If specified, will save 1) .txt file with all significant results and 2) plots for up to 5 labels (with 5 smallest p-values)}

\item{termForSignificance}{the term from your df you want to use for your stats. Must be from the "term" column of plotList[[1]]}
}
\value{
A list object with elements 
    1) a plot for each significant region per your specifications
}
\description{
Make a plot of a bunch of ROI stats from a linear model, per your specifications
}
